fine calculation done

+++ src/fine/FineStrategy.java
@@ -1,6 +1,6 @@
 package fine;
 
 public interface FineStrategy {
-    double calculateFine(int overstayHours);
+    double calculateFine(double overstayHours);
     String getName();
 }
